Fix bug where in-memory contents of deleted file not cleared
This is because `std::filesystem::canonical(path)` requires that path to exist, otherwise it will error and fail.
But the path is now deleted, so it will no longer exist. Converting to weakly_canonical fixes this.

The bug caused memory contents of a file to be preserved even when deleted,
hence when a new file was created of the same name, the contents of
the old file were still kept, and weird behaviour occured

Related to #34