Fix go to definition when table stores result of different module type
Fixes #516

For Go To Definition of a property, we want to find the property in the table it is stored in.
The bug here is that we were using the `definitionModuleType` of the type stored in the property, rather than the table the property is stored in. Whilst the location was of the property itself: this meant that an out-of-sync location was used, causing out-of-bounds failures.