Use virtual module name when resolving string requires
This fixes a bug where the types of a required file would not correctly
update when that file changed. It is because, during string requires,
we use the real file path as the module name.

However, in other usages (e.g., during indexing), we use the virtual
file path if available. This caused an inconsistency between the two.

We fix this by calling "getModuleName" on the returned URI, which
is consistent with indexing.

Ideally in future, we should move towards a state where the
language server always works in the context of real file paths, especially
as string requires become more common. We should instead only
use a sourcemap to resolve virtual requires to real file paths, rather
than the other way round.