Revert "Use Luau's FileUtils rather than filesystem functions"
Turns out this introduces a dependency on MSVC Redistributable which we
previously did not have, causing users who had working language servers
to start crashing. We revert this change to remove that dependency, as
it was depending on windows headers.

It is surprising the MSVC redistributable wasn't originally necessary.
May be in future we will want to statically link against it (if it is
not discouraged?)

This reverts commit 1b593b52747db21d11385805c25c21f261df956a.