Normalise the real path from a source node
Right now, we fail to normalise the file path retrieved from
"getRealPathFromSourceNode". If the path is not canonical (e.g.,
due to usages of '..'), then it will not match the real file path
understood by VSCode. This can cause internal file state to
desync, leading to crashes or odd behaviour.

Fixes #645