Fix regression in ignored globs matching for autoimports
In 00de1bfd7b85908045f69add84f4c84b4eae9eee, we moved the normalisation
of a virtual source node path to the "getRealPathFromSourceNode"
function. This involved applying a call to "weakly_canonical" on the
constructed path.

This call inadvertently caused ignored file detection to break for
virtual file paths. When we call 'isIgnoredFileForAutoImports', we first
get the filepath from 'getRealPathFromSourceNode'. Originally, the
filepath was constructed just by appending the scriptFilePath to the end
of the workspace's root URI, but now it was changed to be normalised as
mentioned above. URIs given by VSCode have lower case drive letters, but
applying this normalisation caused the drive letter to change, e.g. from
c:\ to C:\.

With the file path, we first call 'path.lexically_relative(rootUri)' in
order to relativise the path name with the workspace. However, the first
check in this function is to see whether 'root_name() !=
base.root_name()', and if it does, return the default constructed path
[1]. In our case, the root names no longer matched (due to different
cases), and hence a default constructed path was given.

We fix this by applying an extra normalisation in
"getRealPathFromSourceNode" to first parse it into a Uri, convert it
into a normalised Uri string, then reconstruct a filesystem path from
it.

Fixes #719

[1]: https://en.cppreference.com/w/cpp/filesystem/path/lexically_norma://en.cppreference.com/w/cpp/filesystem/path/lexically_normal