Fix `typeof(x) == "Instance"` refinement
Since Object became the root class on top of Instance, this caused the
typeof refinement for Instance to break.  This is because typeof is
hardcoded to only look for top level "root" classes (i.e., classes with
no parent). But now, Instance has a parent, Object.

However, Luau 0.646 introduced a configuration, `kTypeofRootTag`. If we
attach this tag to the Instance class, then typeof(x) will mark Instance
as a root and refine appropriately.

Fixes #814