Fix crash when hovering over imported type reference
The new documentation on type definitions feature looks up types in a different module scope to find their position.

However, when the old type solver is enabled, we use the module "for autocomplete" for strict datamodel types. But, we did not look up the module for autocomplete when getting documentation for ast node, leading to a failure. In practice, this would return the module computed via "checkSimple" for diagnostics, which does not retain the full type graphs, leading to a bad memory access.