Fix go-to-definition + documentation working on imported type aliases in new solver
This happens when a call to checkSimple (i.e., documentation diagnostics) is done before the go to definition / documentation request.

The underlying cause being that the dependency module was checked with `retainFullTypeGraphs = false`. This means the module scope gets deallocated upon completion, and is no longer available. Even if we `checkStrict` the current module, it doesn't recheck the dependency module.

To fix this, we call `checkStrict` on the dependency module directly. In future, Luau will hopefully provide the alias locations directly even w/o type retention