Optimise didChangeWatchedFiles for large file changes
Right now, didChangeWatchedFiles applies one-by-one on each file. Each time it does this, it marks the file as dirty (when indexing is enabled), then attempts to parse the file.

This is non-optimal if there are a large set of related file changes. Each time, we'll mark as dirty, then parse a single file (which will parse its dependencies). Then, move to the next file, mark it as dirty and repeat. But when we mark it as dirty, we've invalidated all the work we've just done.

We fix this by deferring parsing to after all the changes have been processed, and the dirty list constructed. This prevents unnecessary invalidation.

Also, we now skip ignored files from indexing.