Fix all FFlags being enabled at startup
As it turns out, for the past 2 years luau-lsp has been executing with all fflags enabled.
This is because we are calling 'processFFlags' twice: once for the global command and
once for the subcommand.

The global command does not understand if `--no-flags-enabled` was passed after the subcommand name.
i.e., in the case of 'luau-lsp lsp --no-flags-enabled', it would not see that we don't want to
enable all FFlags by default.

This meant the original processing would enable all fflags, unless --no-flags-enabled was passed twice:
both before and after the `lsp` subcommand.

We fix this by removing the global processFFlags step. Most known use-cases pass the fflag configuration
after the subcommand. We also remove the parent parser registration so that we don't allow fflags before
the subcommand in the first place.

Fixes #1090