Do not clear configStore in workspace/didChangeConfiguration
Right now, we clear our whole configStore when we receive workspace/didChangeConfiguration. Neovim sends
workspace/didChangeConfiguration always as the first message to the client. With our switch to lazy initialization,
this caused the loading of initial configuration to break in Neovim, as we fell back to the default config.

This happens due to the following order of events:

- server initialized
- client sends workspace/didChangeConfiguration (postponed)
- client sends another request (e.g., textDocument/diagnostic) (postponed)
- server requests configuration from client
- client returns configuration, stored in configStore
- configuration is ready, server starts processing postponed messages
- server processes workspace/didChangeConfiguration
    configStore is cleared
- server processes textDocument/diagnostic
    this is the first request for the workspace, so begin lazy initialization
    configStore is empty, so server initializes with default configuration

This did not occur before lazy initialization, as in the old scenario we would immediately setup the workspaces
with configuration after receiving the config.

It does not make sense to clear out the config store in general, as this would mean any messages in between the
config store clearing and the response to the config request would fall back to default config. So, we remove this
step.

Alternatively, we could force initialization of all workspaces upon receiving a 'workspace/didChangeConfiguration'
event. However, this would render lazy initialization useless for Neovim users. Further, clearing the configStore in
general does not make sense as mentioned above