Support cancellation tokens (#1085)
* Support cancellation tokens

* Fix client

* Fix compilation

* Move postponed messages processing to after receiving a response

Right now, there is currently a race condition between the processing loop marking all workspaces as received configuration, and the input loop handling postponed messages. When the processing loop starts processing configuration results, the input loop may have already added new messages to the queue. This means that the processing loop will then process new messages before the input loop gets a chance to add the postponed messages to the queue.

We move the postponed messages handling to after receiveing a response in handleMessage, as we know for a fact that this is when configs will be ready.

* Fix compilation

* Fix changelog structure

* Apply cancellation token for all other requests

* Fix test compilations

* Join message procesor thread on shutdown

If a std::thread is destructed but the thread is still joinable, it will call std::terminate.

Let's ensure we join the thread on shutdown.

We need to update our test cases to also call shutdown upon completion

* Add tests that cancellation is handled

* Add completion test