Fix function parameters autocomplete in calls for new solver
In new solver, the parameters are typed as 'unknown' if they don't have an explicit type annotation.

getParameterExtents treats 'unknown' as an optional type. But note that the new solver would still emit a type error due to mismatching arg counts (see LuauArityMismatchOnUndersaturatedUnknownArguments flag).

Let's manually increment the minimum parameter count when we encounter 'unknown' and treat it as a required type, similar to what this flag does.