Fix crash when accessing cancellationTokens map concurrently
We are currently inserting into the cancellationTokens map without holding a mutex lock. Note that the processor thread could erase a cancellationToken from the map at the same time upon request processing completion. The clearing is correctly under a mutex lock.

Similarly, when looking up the cancellationToken inside of onRequest, we are not doing this under a mutex lock.

We add a lock in both locations