Move plugin info mutation into `updateSourceNodeMap`
Right now, it's in `handleSourcemapUpdate`. However handleSourcemapUpdate is called twice in the old solver - one for the normal typechecker and one for autocomplete. This means we end up calling mutate twice, and perform the addition twice.

Let's move it to `updateSourceNodeMap` instead, where it is only resolved once.

Whilst here, move write paths out of the try-catch, as that is not what we are testing against. Make sure we are resetting the state in the catch block.