Revert "Update to latest types dump (#1283)"
This reverts commit ef1234c3a40b6c73d9c522d9f98b25f236b3c5cf.

This definitions update deprecates a lot of methods and replaces them with variants with an "Async" suffix. However, these variants are currently not enabled on all clients, so this throws an error in actual Studio. Let's revert the update until the new methods are actually available to use in Roblox.