feat!: add LoadLibrary support
BREAKING CHANGE: Okay so this isn't really a breaking change but more of a warning that this commit shouldn't be released.

Currently the CLI loads wayyyy too slowly to be a viable replacement and there's a few wack things here and there as well. Should you choose to use this version, please note that things will be slow and some things may break.

The next steps to fix these issues are to maintain a local map of all libraries - probably in a json file that is generated asynchronously to the lsp indexing files - most likely done as a VSC task and then yielding until that map is generated to load libraries. Since it should be a task (could be a script on 2nd thought) there should also be a feature to update the map to reference new libraries.

Hopefully we won't have to replace `LoadLibrary` calls with `require` calls under the hood and instead we can just typecast the variable instead.