Only index alias files if they are external to the repository.
If we have an alias that is pointing to within the repo (e.g., @game/), we'll end up indexing the whole repo twice (because the logic in traverse directory always assumes a unique file).

This won't cause any fundamental issue in practice, because Frontend should deduplicate, however it leads to an double in size of `moduleNames`, and we're also more likely to hit the `maxFiles` limit unnecessarily.