Always mark externally changed files as dirty
Right now, if an externally changed file is ignored, we don't mark it as
dirty. This means that internally Luau does not realise the file
contents have changed, and will continue typechecking with stale
contents until the file is explicitly opened in the editor.

We change it so that we always mark a file as dirty regardless of
ignored state.

There is a potential concern that "dirtyFiles" will now include ignored
files and we will parse them, but this is not an issue in practice:

- If this ignored file was never indexed, dirtyFiles will not extend at
  all because there is no SourceModule registered
- If this ignored file was indexed before (i.e., there is a non-ignored
  file that has a [transitive] require to this file), then dirtyFiles
will extend. However, there is no point trying to filter the dirtyFiles
for non-ignored files, as there is guaranteed to be a non-ignored file
in the dirtyFiles closure that [transitively] depends on all these
ignored files, so they will get re-indexed anyways.

Fixes #1144